Add PanelViewToggle and modularized code
