+++ src/app/ComponentView.jsx
@@ -1,0 +1,11 @@
+import React, { useEffect } from 'react';
+
+const ComponentView = () => {
+  return (
+    <>
+      <p>ComponentView</p>
+    </>
+  );
+};
+
+export default ComponentView;
